requirements.txt
appdirs==1.4.3
packaging==16.8
paho-mqtt==1.2
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
